BUG: Fix decompressing all zlib data for NIDS
Code was only expanding the buffer by 10000 bytes as it read through.
This resulted in blowing the array bounds when the last, noncompressed
chunk of data was larger than this. This was easily broken by products
with a single compressed chunk, followed by all uncompressed data. A
representative data file and smoke test is included.